Fix non-debug compilation (#52)
* Fix AssertFO in non-debug versions


* Fix formatting


* Fix float-double cast warnings

+++ NeoML/src/Dnn/Layers/ModelWrapperLayer.cpp
@@ -319,10 +319,10 @@ bool CDnnModelWrapper::classify( CClassificationResult& result ) const
 		NeoAssert(resultBlob->GetObjectSize() == 1);
 
 		// Use a sigmoid function to estimate probabilities
-		const float zeroClassProb = 1 / (1 + exp(resultBlob->GetData().GetValue()));
+		const double zeroClassProb = 1 / (1 + exp(resultBlob->GetData().GetValue()));
 		result.Probabilities[0] = CClassificationProbability(zeroClassProb);
 		result.Probabilities[1] = CClassificationProbability(1 - zeroClassProb);
-		result.PreferredClass = zeroClassProb >= 0.5f ? 0 : 1;
+		result.PreferredClass = zeroClassProb >= 0.5 ? 0 : 1;
 	} else {
 		NeoAssert(resultBlob->GetObjectSize() == ClassCount);
